backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
pdfplumber==0.11.0
pandas==2.2.2
openpyxl==3.1.5
requests==2.32.3
